Scholium/requirements.txt
langgraph == 0.2.66
langchain-openai == 0.2.9
langgraph-checkpoint == 2.0.10

pinecone[grpc]
pinecone == 5.4.2
python-dotenv == 1.0.1

fastapi[standard] == 0.115.0
pydantic>=2.7.0,<3.0.0
uvicorn == 0.34.0
copilotkit == 0.1.34

ipython == 8.32.0
openai == 1.63.2
